commit 9

+++ blog/views.py
@@ -5,9 +5,11 @@ from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from rest_framework.views import APIView
 import requests
-
-from .models import Article
-from .serializers import UserSerializers, ArticleSerializers
+from rest_framework.authentication import TokenAuthentication
+from .models import Article, Comment
+from .serializers import UserSerializers, ArticleSerializers, CommentSerializers
+from rest_framework.permissions import IsAuthenticated
+from .permissions import IsUserOrReadOnly
 
 # Create your views here.
 
@@ -55,19 +57,22 @@ class ArticleDetailView(APIView):
 
 
 class AddArticleView(APIView):
+    permission_classes = [IsAuthenticated]
+
     def post(self, request):
-        serializer = ArticleSerializers(data=request.data)
+        serializer = ArticleSerializers(data=request.data, context={'request': request})
         if serializer.is_valid():
-            instance = serializer.save()
-            instance.published = True
-            instance.save()
+            serializer.save()
             return Response({"response": "Done"}, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 class ArticleUpdateView(APIView):
+    permission_classes = [IsUserOrReadOnly]
+
     def put(self, request, pk):
         instance = Article.objects.get(id=pk)
+        self.check_object_permissions(request, instance)
         serializer = ArticleSerializers(instance=instance, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
@@ -76,5 +81,19 @@ class ArticleUpdateView(APIView):
 
     def delete(self, request, pk):
         instance = Article.objects.get(id=pk)
+        self.check_object_permissions(request, instance)
         instance.delete()
         return Response({"response": "Deleted"}, status=status.HTTP_200_OK)
+
+
+class CheckToken(APIView):
+    def get(self, request):
+        user = request.user
+        return Response({'user': user.username}, status=status.HTTP_200_OK)
+
+
+class GetArticlesCommentView(APIView):
+    def get(self, request, pk):
+        comments = Article.objects.get(id=pk).comments.all()
+        ser = CommentSerializers(instance=comments, many=True)
+        return Response(ser.data, status=status.HTTP_200_OK)
